requirements.txt
pandas
numpy
python-dotenv
gradio
openai
sentence-transformers
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-chroma
langchain-huggingface
